requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
chromadb
sentence-transformers
google-cloud-aiplatform
google-cloud-vision
google-cloud-speech
google-cloud-texttospeech
ffmpeg-python
beautifulsoup4
requests
python-multipart